Create blind-auction-start
## Blind Auction

Click "Open Preview" above to see this file rendered with the markdown.

# Instructions

The objective is to write a program that will collect the names and bids of different people. The program should ask for each bidder's name and their bid individually. 

```
Welcome to the secret auction program. 
What is your name?: Angela
```
```
What's your bid?: $123
```
```
Are there any other bidders? Type 'yes' or 'no'.
yes

```
If there are other bidders, the screen should clear, so you can pass your phone to the next person. If there are no more bidders, then the program should display the name of the winner and their winning bid. 

```
The winner is Elon with a bid of $55000000000
```

Use your knowledge of Python dictionaries and loops to solve this challenge. 


# My console doesn't clear!

This will happen if you’re using an IDE other than replit (e.g., VSCode, PyCharm etc). Similar to how we used the "random" module previously, in this project we will use the "replit" module. The `clear()` function is available here via the replit module without any extra configuration. 

**I’ll cover how to use PyCharm and import modules on Day 15**. That said, you can write your own `clear()` function or configure your IDE like so: 

[Udemy Q&A Answer](https://www.udemy.com/course/100-days-of-code/learn/lecture/19279420#questions/16084076)


# Solution

[https://replit.com/@appbrewery/blind-auction-completed](https://replit.com/@appbrewery/blind-auction-completed?v=1)